# Speed up date_histogram by precomputing ranges (backport of #61467)